requirements.txt
yfinance
pandas
scikit-learn
google-generativeai
streamlit
plotly
numpy
requests
diskcache
